Revert "fix(ext/console): fix inspecting iterators error. (#20720)"
This reverts commit 0209f7b46954d1b7bf923b4191e5a356ec09622c.
