+++ ext/console/01_console.js
@@ -136,7 +136,6 @@ const {
   WeakSetPrototypeHas,
   isNaN,
 } = primordials;
-import { previewEntries } from "ext:deno_node/internal_binding/util.ts";
 
 let noColor = () => false;
 
@@ -1494,7 +1493,9 @@ function getIteratorBraces(type, tag) {
 
 const iteratorRegExp = new SafeRegExp(" Iterator] {$");
 function formatIterator(braces, ctx, value, recurseTimes) {
-  const { 0: entries, 1: isKeyValue } = previewEntries(value, true);
+  // TODO(wafuwafu13): Implement
+  // const { 0: entries, 1: isKeyValue } = previewEntries(value, true);
+  const { 0: entries, 1: isKeyValue } = value;
   if (isKeyValue) {
     // Mark entry iterators as such.
     braces[0] = StringPrototypeReplace(
@@ -1703,12 +1704,16 @@ function formatWeakCollection(ctx) {
 }
 
 function formatWeakSet(ctx, value, recurseTimes) {
-  const entries = previewEntries(value);
+  // TODO(wafuwafu13): Implement
+  // const entries = previewEntries(value);
+  const entries = value;
   return formatSetIterInner(ctx, recurseTimes, entries, kWeak);
 }
 
 function formatWeakMap(ctx, value, recurseTimes) {
-  const entries = previewEntries(value);
+  // TODO(wafuwafu13): Implement
+  // const entries = previewEntries(value);
+  const entries = value;
   return formatMapIterInner(ctx, recurseTimes, entries, kWeak);
 }
