+++ ext/node/polyfills/internal/console/constructor.mjs
@@ -17,7 +17,17 @@ import {
   validateInteger,
   validateObject,
 } from "ext:deno_node/internal/validators.mjs";
-import { previewEntries } from "ext:deno_node/internal_binding/util.ts";
+const previewEntries = (iter, isKeyValue) => {
+  if (isKeyValue) {
+    const arr = [...iter];
+    if (Array.isArray(arr[0]) && arr[0].length === 2) {
+      return [[].concat(...arr), true];
+    }
+    return [arr, false];
+  } else {
+    return [...iter];
+  }
+};
 import { Buffer } from "node:buffer";
 const { isBuffer } = Buffer;
 import {
@@ -465,6 +475,7 @@ const consoleMethods = {
 
   // https://console.spec.whatwg.org/#table
   table(tabularData, properties) {
+    console.log("tabularData", tabularData);
     if (properties !== undefined) {
       validateArray(properties, "properties");
     }
